Move argparse checkers to io_functions.py
